+++ agents/sac/__init__.py
@@ -1,2 +1,2 @@
-from .rollout import *
+from .networks import *
 from .sac import *
