+++ agents/sac/sac.py
@@ -1,174 +1,85 @@
+import collections
 import copy
+import os
+import time
+from dataclasses import dataclass
 
+import gym
+import numpy as np
 import torch
-import torch.nn as nn
 import torch.nn.functional as F
-from torch.distributions import Normal
-from agents.utils.experiment import unpack_batch
 
-LOG_SIG_MAX = 2
-LOG_SIG_MIN = -20
-epsilon = 1e-6
+from agents.utils import HyperParameters, NStepTracer, generate_gif, unpack_batch
 
 
-# Initialize Policy weights
-def weights_init_(m):
-    if isinstance(m, nn.Linear):
-        torch.nn.init.xavier_uniform_(m.weight, gain=1)
-        torch.nn.init.constant_(m.bias, 0)
+@dataclass
+class SACHP(HyperParameters):
+    ALPHA: float = 0.015
+    LOG_SIG_MAX: int = 2
+    LOG_SIG_MIN: int = -20
+    EPSILON: float= 1e-6
 
 
-class QNetwork(nn.Module):
-    def __init__(self, num_inputs, num_actions, hidden_dim=256):
-        super(QNetwork, self).__init__()
+def data_func(
+    pi,
+    device,
+    queue_m,
+    finish_event_m,
+    gif_req_m,
+    hp
+):
+    env = gym.make(hp.ENV_NAME)
+    tracer = NStepTracer(n=hp.REWARD_STEPS, gamma=hp.GAMMA)
 
-        # Q1 architecture
-        self.q1 = nn.Sequential(
-            nn.Linear(num_inputs + num_actions, hidden_dim),
-            nn.ReLU(),
-            nn.Linear(hidden_dim, hidden_dim),
-            nn.ReLU(),
-            nn.Linear(hidden_dim, 1)
-        )
+    with torch.no_grad():
+        while not finish_event_m.is_set():
+            # Check for generate gif request
+            gif_idx = -1
+            with gif_req_m.get_lock():
+                if gif_req_m.value != -1:
+                    gif_idx = gif_req_m.value
+                    gif_req_m.value = -1
+            if gif_idx != -1:
+                path = os.path.join(hp.SAVE_PATH, f"gifs/{gif_idx:09d}.gif")
+                generate_gif(env=env, filepath=path,
+                             pi=copy.deepcopy(pi),
+                             max_episode_steps=1000, device=device)
 
-        # Q2 architecture
-        self.q2 = nn.Sequential(
-            nn.Linear(num_inputs + num_actions, hidden_dim),
-            nn.ReLU(),
-            nn.Linear(hidden_dim, hidden_dim),
-            nn.ReLU(),
-            nn.Linear(hidden_dim, 1)
-        )
+            done = False
+            s = env.reset()
+            ep_steps = 0
+            ep_rw = 0
+            st_time = time.perf_counter()
+            while not done:
+                # Step the environment
+                s_v = torch.Tensor(s).to(device)
+                a = pi.get_action(s_v)
+                s_next, r, done, info = env.step(a)
+                ep_steps += 1
+                ep_rw += r
 
-        self.apply(weights_init_)
+                # Trace NStep rewards and add to mp queue
+                tracer.add(s, a, r, done)
+                while tracer:
+                    queue_m.put(tracer.pop())
 
-    def forward(self, state, action):
-        xu = torch.cat([state.clone(), action.clone()], 1)
-        x1 = self.q1(xu)
-        x2 = self.q2(xu)
-        return x1, x2
+                if done:
+                    info['fps'] = ep_steps / (time.perf_counter() - st_time)
+                    info['ep_steps'] = ep_steps
+                    info['ep_rw'] = ep_rw
+                    queue_m.put(info)
 
-
-class GaussianPolicy(nn.Module):
-    def __init__(self, num_inputs, num_actions,
-                 log_sig_min, log_sig_max,
-                 epsilon, hidden_dim=256, action_space=None):
-        super(GaussianPolicy, self).__init__()
-        self.log_sig_min = log_sig_min
-        self.log_sig_max = log_sig_max
-        self.epsilon = epsilon
-        self.linear1 = nn.Linear(num_inputs, hidden_dim)
-        self.linear2 = nn.Linear(hidden_dim, hidden_dim)
-
-        self.mean_linear = nn.Linear(hidden_dim, num_actions)
-        self.log_std_linear = nn.Linear(hidden_dim, num_actions)
-
-        self.apply(weights_init_)
-
-        # action rescaling
-        if action_space is None:
-            self.action_scale = torch.tensor(1.)
-            self.action_bias = torch.tensor(0.)
-        else:
-            self.action_scale = torch.FloatTensor(
-                (action_space.high - action_space.low) / 2.)
-            self.action_bias = torch.FloatTensor(
-                (action_space.high + action_space.low) / 2.)
-
-    def forward(self, state):
-        x1 = F.relu(self.linear1(state))
-        x2 = F.relu(self.linear2(x1))
-        mean = self.mean_linear(x2)
-        log_std = self.log_std_linear(x2)
-        log_std = torch.clamp(log_std, min=self.log_sig_min,
-                              max=self.log_sig_max)
-        return mean, log_std
-
-    def sample(self, state):
-        mean, log_std = self.forward(state)
-        std = log_std.exp()
-        normal = Normal(mean, std)
-        # for reparameterization trick (mean + std * N(0,1))
-        x_t = normal.rsample()
-        y_t = torch.tanh(x_t)
-        action = y_t * self.action_scale + self.action_bias
-        log_prob = normal.log_prob(x_t)
-        # Enforcing Action Bound
-        log_action = torch.log(
-            self.action_scale * (1 - y_t.pow(2)) + self.epsilon
-        )
-        log_prob = log_prob - log_action
-        log_prob = log_prob.sum(1, keepdim=True)
-        mean = torch.tanh(mean) * self.action_scale + self.action_bias
-        return action, log_prob, mean
-
-    def get_action(self, state):
-        action, _, _ = self.sample(state.unsqueeze(0))
-        return action.detach().cpu().numpy()[0]
-
-    def to(self, device):
-        self.action_scale = self.action_scale.to(device)
-        self.action_bias = self.action_bias.to(device)
-        return super(GaussianPolicy, self).to(device)
-
-
-class TargetNet:
-    """
-    Wrapper around model which provides copy of it instead of trained weights
-    """
-
-    def __init__(self, model):
-        self.model = model
-        self.target_model = copy.deepcopy(model)
-
-    def sync(self, alpha):
-        """
-        Blend params of target net with params from the model
-        :param alpha:
-        """
-        assert isinstance(alpha, float)
-        assert 0.0 < alpha <= 1.0
-        state = self.model.state_dict()
-        tgt_state = self.target_model.state_dict()
-        for k, v in state.items():
-            tgt_state[k] = tgt_state[k] * alpha + (1 - alpha) * v
-        self.target_model.load_state_dict(tgt_state)
-
-
-class TargetActor(TargetNet):
-    def __call__(self, S):
-        return self.target_model(S)
-
-
-class TargetCritic(TargetNet):
-    def __call__(self, S, A):
-        return self.target_model(S, A)
+                # Set state for next step
+                s = s_next
 
 
 def loss_sac(alpha, gamma, batch, crt_net, act_net,
-             tgt_crt_net, cuda=False, cuda_async=True):
+             tgt_crt_net, device):
 
     state_batch, action_batch, reward_batch,\
-        mask_batch, next_state_batch = unpack_batch(batch)
+        mask_batch, next_state_batch = unpack_batch(batch, device)
 
-    state_batch = torch.tensor(state_batch, dtype=torch.float32)
-    next_state_batch = torch.tensor(next_state_batch, dtype=torch.float32)
-    action_batch = torch.tensor(action_batch, dtype=torch.float32)
-    reward_batch = torch.tensor(reward_batch, dtype=torch.float32).unsqueeze(1)
-    mask_batch = torch.BoolTensor(mask_batch)
-
-    if cuda:
-        state_batch = state_batch.cuda(non_blocking=cuda_async)
-        next_state_batch = next_state_batch.cuda(non_blocking=cuda_async)
-        action_batch = action_batch.cuda(non_blocking=cuda_async)
-        reward_batch = reward_batch.cuda(non_blocking=cuda_async)
-        mask_batch = mask_batch.cuda(non_blocking=cuda_async)
-    else:
-        state_batch = torch.FloatTensor(state_batch)
-        next_state_batch = torch.FloatTensor(next_state_batch)
-        action_batch = torch.FloatTensor(action_batch)
-        reward_batch = torch.FloatTensor(reward_batch)
-        mask_batch = torch.BoolTensor(mask_batch)
+    reward_batch = reward_batch.unsqueeze_(1)
 
     with torch.no_grad():
         next_state_action, next_state_log_pi, _ = act_net.sample(
@@ -206,3 +117,35 @@ def loss_sac(alpha, gamma, batch, crt_net, act_net,
     policy_loss = policy_loss.mean()
 
     return policy_loss, qf1_loss, qf2_loss, log_pi
+
+
+def save_checkpoint(
+    experiment: str,
+    agent: str,
+    pi,
+    Q,
+    pi_opt,
+    Q_opt,
+    alpha,
+    n_samples,
+    n_grads,
+    n_episodes,
+    device,
+    checkpoint_path: str
+):
+    checkpoint = {
+        "name": experiment,
+        "agent": agent,
+        "pi_state_dict": pi.state_dict(),
+        "Q_state_dict": Q.state_dict(),
+        "pi_opt_state_dict": pi_opt.state_dict(),
+        "Q_opt_state_dict": Q_opt.state_dict(),
+        "alpha": alpha,
+        "n_samples": n_samples,
+        "n_grads": n_grads,
+        "n_episodes": n_episodes,
+        "device": device
+    }
+    filename = os.path.join(
+        checkpoint_path, "checkpoint_{:09}.pth".format(n_grads))
+    torch.save(checkpoint, filename)
